requirements.txt
asgiref==3.7.2
boto3==1.34.16
botocore==1.34.16
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==5.0.1
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
drf-social-oauth2==2.1.3
ecdsa==0.18.0
idna==3.6
jmespath==1.0.1
jwcrypto==1.5.1
oauthlib==3.2.2
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.0.7
wrapt==1.16.0
